Use postgres instead of sqlite3 for database backend.
Adds the `pg` gem as a dependency and defaults to using postgres
for development, test, and production configurations of the application.

You must have DATABASE_URL in your .env set to some value (for
development and testing this can be `postgres://localhost`).

After pulling this commit you will need to run
`bundle exec rake db:setup` which will create the development and test
databases and apply all of the migrations.